Rectify 2.5: capability-tied visibility — durable session-scoped exploration authority (#4684)
- Extract _ReopenedLaunchAuthority, _safe_submit_failure_reason, and
  _ExplorationLaunchAuthorityStore out of exploration_context.py into a new
  sibling module exploration_context_durable.py, to stay under the 1100-line
  ceiling while adding new durable-authority machinery. One-way dependency
  (durable -> exploration_context); no circular import.
- Add OwnerBoundExplorationContextStore.lease_for_capability() accessor so
  callers can resolve the in-memory lease minted for a just-bound capability.
- Add bind_session_scoped_durable() free function in the new module: binds
  the in-memory capability via bind_session_scoped(), then writes an
  HMAC-signed 0600 durable authority file recording the same capability,
  mirroring bind_launch's durability guarantee for the session-scoped path.
- Wire tools_exploration.py's enable_exploration to call
  bind_session_scoped_durable(...) with a per-session authority_home under
  ctx.temp_dir, instead of the bare in-memory bind_session_scoped().
- Fix grant/revoke asymmetry: enable_exploration granted FastMCP tag
  visibility via ctx.enable_components() but only ever revoked the in-memory
  lease via store.cleanup_session() on failure paths — never the tag itself.
  Add await ctx.disable_components(tags={"exploration"}) to the finally
  block so a later failure after a successful grant always revokes the tag.
- Register both durable authority writers (bind_launch,
  bind_session_scoped_durable) in DURABLE_ARTIFACT_WRITERS
  (core/types/_type_constants.py).
- Add tests/contracts/test_exploration_grant_revoke_symmetry.py: an AST
  check that every ctx.enable_components(...) call site in
  tools_exploration.py has a reachable ctx.disable_components(...) in the
  same function, plus a membership check that both durable writers are
  registered. Deliberately does NOT add pipeline/exploration_context*.py to
  test_durable_artifact_writers_guard.py's _SCOPED_MODULES — that guard's
  docstring explicitly rejects codebase-wide scope expansion beyond hook
  artifacts; the membership check above is the plan's actual specified home
  for this assertion.
- Update tests/server/test_enable_exploration.py and
  test_enable_exploration_failure_codes.py: ctx fixtures now set
  disable_components = AsyncMock() and assert it is awaited with
  tags={"exploration"} on both the cancellation path and the
  enable_components-failure path.
- Bump tests/arch/test_subpackage_isolation.py's pipeline file-count
  exemption from 18 to 19 for the new sibling module.

test_check: 37859 passed, 613 skipped, 27 xfailed.
